# openapi.yaml
openapi: 3.0.0
info:
  title: User API
  version: 1.0.0
  description: API для управления пользователем, его клиентами, продуктами и заказами
servers:
  - url: http://localhost:5000
    description: Локальный сервер разработки
...
paths:
  /user/register:
    post:
      tags:
        - user
      summary: Регистрация нового пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                firstname:
                  type: string
                lastname:
                  type: string
                phone:
                  type: string
                password:
                  type: string
                  format: password
      responses:
        '201':
          description: Пользователь успешно зарегистрирован
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User registered successfully"
        '400':
          description: Ошибка валидации данных
        '409':
          description: Пользователь с таким email уже существует
  /user/login:
    post:
      tags:
        - user
      summary: Вход пользователя в систему
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Login successful"
                  access_token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '400':
          description: Ошибка валидации данных
        '401':
          description: Неверный email или пароль
  /user/me:
    get:
      tags:
       - user
      summary: Информация о пользователе
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Успешное получение данных пользователя
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  email:
                    type: string
                  first_name:
                    type: string
                  last_name:
                    type: string
                  phone:
                    type: string
                example:
                  id: 1
                  email: user@example.com
                  first_name: Иван
                  last_name: Иванов
                  phone: "+79161234567"
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: "User not found"
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT